begin refactoring
