+++ src/lib/diff/myers.test.ts
@@ -1,20 +1,11 @@
 import { test, expect } from 'vitest'
-import { myersDiffClean } from "./myers";
+import { myersDiffRaw } from "./myers";
 import { DiffChunkOp } from './diff.d';
 
 test("Myer's diff correctness checks.", async () => {
-
-  let checkChunks = (A: string, B: string, len: number) => {
-    let chunks = myersDiffClean(A, B);
-    expect(chunks.length).toBe(len)
-  }
-
-  checkChunks("stuff", "stutff", 3);
-  checkChunks("stufft", "stuff", 2);
-  checkChunks("abcd", "zbczz", 5);
-  const sames = myersDiffClean("The cat in the hat.", "The dig in the rig.").filter(c => c.op == DiffChunkOp.Equal);
+  const sames = myersDiffRaw("The cat in the hat.", "The dog in the log.").filter(c => c.op == DiffChunkOp.Equal);
   expect(sames.length).toBe(3);
   expect(sames[0].data).toBe('The ');
-  expect(sames[2].data).toBe('.');
+  expect(sames.at(-1)?.data).toBe('.');
 })
